dddd

+++ src/clases/Filtro.java
@@ -1,5 +1,5 @@
 package clases;
 
 public interface Filtro {
-public double ejecutar(int revoluciones);
+public double ejecutar(double revoluciones);
 }
